separate lyrics from songs
